requirements.txt
fastapi
uvicorn
whisper
ffmpeg-python
openai
python-dotenv
requests
